restructure file system and imports
