+++ src/config/pkg.mjs
