+++ src/core/handlers/auth/index.mjs
@@ -1,10 +1,10 @@
 import crypto from "crypto";
 
-import { hashPassword, randomToken } from "$/core/utils/auth.mjs";
-import logger from "$/core/utils/logger.mjs";
-import { sendMail } from "$/core/utils/mailer.mjs";
-import env from "$/core/config/env.mjs";
-import prisma from "$/core/services/database.mjs";
+import { prisma } from "$/services/database.mjs";
+import { hashPassword, randomToken } from "$/utils/auth.mjs";
+import logger from "$/utils/logger.mjs";
+import { sendMail } from "$/utils/mailer.mjs";
+import env from "$/config/env.mjs";
 
 const { APP_URL, AUTH_SESSION_COOKIE_NAME, COOKIE_OPTS, APP_NAME } = env();
