+++ src/core/middlewares/auth.mjs
@@ -2,8 +2,8 @@ import {
   getSessionWithUser,
   getOrCreateSingletonGuestUser,
   createSession,
-} from "$/core/utils/auth.mjs";
-import env from "$/core/config/env.mjs";
+} from "$/utils/auth.mjs";
+import env from "$/config/env.mjs";
 
 const {
   AUTH_SESSION_COOKIE_NAME,
