+++ src/core/middlewares/useJSX.mjs
@@ -4,7 +4,7 @@ import path from "path";
 import React from "react";
 import { renderToString } from "react-dom/server";
 
-import env from "$/core/config/env.mjs";
+import env from "$/config/env.mjs";
 
 const { NODE_ENV } = env();
