+++ src/plugins/blog/handlers/index.mjs
@@ -1,17 +1,17 @@
 import path from "path";
 
-import logger from "$/core/utils/logger.mjs";
-import prisma from "$/core/services/database.mjs";
+import { prisma } from "$/services/database.mjs";
 import {
   getGitManager,
   getOrInitGitManager,
   disposeGitManager,
-} from "$/core/libs/git/registry.mjs";
-import FileManager from "$/core/libs/fs.mjs";
+} from "$/libs/git/registry.mjs";
+import FileManager from "$/libs/fs.mjs";
+import logger from "$/utils/logger.mjs";
 import {
   ensureProjectAccess,
   ProjectAccessError,
-} from "$/core/utils/projectAccess.mjs";
+} from "$/utils/projectAccess.mjs";
 
 const PAPERTRAIL_BOARD_SELECT = {
   id: true,
