+++ tests/projects-shares.test.mjs
@@ -1,7 +1,7 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 
-import prisma from "$/core/services/database.mjs";
+import { prisma } from "$/services/database.mjs";
 import * as shares from "$/core/handlers/projects/shares.mjs";
 
 function buildRes() {
